move to entity-item module
