+++ src/Umbraco.Web.UI.Client/src/packages/core/entity-item/data-api-item-get-request-controller/index.ts
@@ -1,0 +1,1 @@
+export * from './data-api-item-get-request.controller.js';
