+++ src/Umbraco.Web.UI.Client/src/packages/core/entity-item/data-api-item-get-request-controller/types.ts
@@ -1,0 +1,6 @@
+import type { UmbDataApiResponse } from '@umbraco-cms/backoffice/resources';
+
+export interface UmbDataApiItemGetRequestControllerArgs<ResponseModelType extends UmbDataApiResponse> {
+	api: (args: { uniques: Array<string> }) => Promise<ResponseModelType>;
+	uniques: Array<string>;
+}
