+++ src/Umbraco.Web.UI.Client/src/packages/core/entity-item/index.ts
@@ -1,1 +1,2 @@
+export * from './data-api-item-get-request-controller/index.js';
 export * from './entity-item-ref/index.js';
