+++ src/Umbraco.Web.UI.Client/src/packages/core/entity-item/types.ts
@@ -1,4 +1,5 @@
 import type { UmbNamedEntityModel } from '@umbraco-cms/backoffice/entity';
+export type * from './data-api-item-get-request-controller/types.js';
 
 export interface UmbDefaultItemModel extends UmbNamedEntityModel {
 	icon?: string;
