+++ src/Umbraco.Web.UI.Client/src/packages/core/resources/data-api/types.ts
@@ -1,0 +1,3 @@
+export interface UmbDataApiResponse<ResponseType extends { data: unknown } = { data: unknown }> {
+	data: ResponseType['data'];
+}
