+++ src/Umbraco.Web.UI.Client/src/packages/media/media/repository/item/media-item.server.data-source.ts
@@ -4,16 +4,8 @@ import type { MediaItemResponseModel } from '@umbraco-cms/backoffice/external/ba
 import { MediaService } from '@umbraco-cms/backoffice/external/backend-api';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbItemServerDataSourceBase } from '@umbraco-cms/backoffice/repository';
-import {
-	UmbError,
-	type UmbApiError,
-	type UmbCancelError,
-	type UmbDataApiResponse,
-} from '@umbraco-cms/backoffice/resources';
-import { batchTryExecute, tryExecute } from '@umbraco-cms/backoffice/resources';
-import { batchArray } from '@umbraco-cms/backoffice/utils';
-import { UmbControllerBase } from '@umbraco-cms/backoffice/class-api';
-import { umbPeekError } from '@umbraco-cms/backoffice/notification';
+import { tryExecute } from '@umbraco-cms/backoffice/resources';
+import { UmbDataApiItemGetRequestController } from '@umbraco-cms/backoffice/entity-item';
 
 /**
  * A data source for Media items that fetches data from the server
@@ -95,58 +87,3 @@ const mapper = (item: MediaItemResponseModel): UmbMediaItemModel => {
 		}),
 	};
 };
-
-interface UmbDataApiItemGetRequestControllerArgs<ResponseModelType extends UmbDataApiResponse> {
-	api: (args: { uniques: Array<string> }) => Promise<ResponseModelType>;
-	uniques: Array<string>;
-}
-class UmbDataApiItemGetRequestController<ResponseModelType extends UmbDataApiResponse> extends UmbControllerBase {
-	#apiCallback: (args: { uniques: Array<string> }) => Promise<ResponseModelType>;
-	#uniques: Array<string>;
-	#batchSize: number = 1;
-
-	constructor(host: UmbControllerHost, args: UmbDataApiItemGetRequestControllerArgs<ResponseModelType>) {
-		super(host);
-		this.#apiCallback = args.api;
-		this.#uniques = args.uniques;
-	}
-
-	async request() {
-		if (!this.#uniques) throw new Error('Uniques are missing');
-
-		let data: ResponseModelType['data'] | undefined;
-		let error: UmbError | UmbApiError | UmbCancelError | Error | undefined;
-
-		if (this.#uniques.length > this.#batchSize) {
-			const chunks = batchArray<string>(this.#uniques, this.#batchSize);
-			const results = await batchTryExecute(this, chunks, (chunk) => this.#apiCallback({ uniques: chunk }));
-
-			const errors = results.filter((promiseResult) => promiseResult.status === 'rejected');
-
-			if (errors.length > 0) {
-				error = await this.#getAndHandleErrorResult(errors);
-			}
-
-			data = results
-				.filter((promiseResult) => promiseResult.status === 'fulfilled')
-				.flatMap((promiseResult) => promiseResult.value.data);
-		} else {
-			const result = await tryExecute(this, this.#apiCallback({ uniques: this.#uniques }));
-			data = result.data;
-			error = result.error;
-		}
-
-		return { data, error };
-	}
-
-	async #getAndHandleErrorResult(errors: Array<PromiseRejectedResult>) {
-		// TODO: We currently expect all the errors to be the same, but we should handle this better in the future.
-		const error = errors[0];
-		await umbPeekError(this, {
-			headline: 'Error fetching items',
-			message: 'An error occurred while fetching items.',
-		});
-
-		return new UmbError(error.reason);
-	}
-}
